requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dacite==1.6.0
defusedxml==0.7.1
Django==4.0.5
django-filter==21.1
django-pytest==0.2.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.14.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
